+++ packages/config/src/index.ts
@@ -42,6 +42,21 @@ function getNormalizedChangelogOption(
   return thing;
 }
 
+function getNormalizedCommitOption(
+  thing: boolean | readonly [string, any] | string
+): Config["commit"] {
+  if (thing === false) {
+    return false;
+  }
+  if (thing === true) {
+    return ["@changesets/cli/commit", { skipCI: "version" }];
+  }
+  if (typeof thing === "string") {
+    return [thing, null];
+  }
+  return thing;
+}
+
 function getUnmatchedPatterns(
   listOfPackageNamesOrGlob: readonly string[],
   pkgNames: readonly string[]
@@ -126,13 +141,22 @@ export let parse = (json: WrittenConfig, packages: Packages): Config => {
     );
   }
 
-  if (json.commit !== undefined && typeof json.commit !== "boolean") {
+  if (
+    json.commit !== undefined &&
+    typeof json.commit !== "boolean" &&
+    typeof json.commit !== "string" &&
+    !(
+      isArray(json.commit) &&
+      json.commit.length === 2 &&
+      typeof json.commit[0] === "string"
+    )
+  ) {
     messages.push(
       `The \`commit\` option is set as ${JSON.stringify(
         json.commit,
         null,
         2
-      )} when the only valid values are undefined or a boolean`
+      )} when the only valid values are undefined or a boolean or a module path (e.g. "@changesets/cli/commit" or "./some-module") or a tuple with a module path and config for the commit message generator (e.g. ["@changesets/cli/commit", { "skipCI": "version" }])`
     );
   }
   if (json.baseBranch !== undefined && typeof json.baseBranch !== "string") {
@@ -351,8 +375,9 @@ export let parse = (json: WrittenConfig, packages: Packages): Config => {
       normalizedAccess === undefined
         ? defaultWrittenConfig.access
         : normalizedAccess,
-    commit:
-      json.commit === undefined ? defaultWrittenConfig.commit : json.commit,
+    commit: getNormalizedCommitOption(
+      json.commit === undefined ? defaultWrittenConfig.commit : json.commit
+    ),
     fixed,
     linked,
     baseBranch: