+++ packages/write/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @changesets/write
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [[`c87eba6`](https://github.com/changesets/changesets/commit/c87eba6f80a34563b7382f87472c29f6dafb546c)]:
+  - @changesets/types@5.0.0
+
 ## 0.1.7
 
 ### Patch Changes